cost-optimization/gke-vpa-recommendations/metrics-exporter/requirements.txt
google-cloud-bigquery-storage
google-cloud-bigquery
pandas
pandas_gbq
tqdm
aiohttp
psutil